implement win rate percentage circle for each faction
