+++ index.html
@@ -5,7 +5,6 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <meta http-equiv="X-UA-Compatible" content="ie=edge" />
     <link rel="stylesheet" href="./src/styles/index.css" />
-    <script src="https://d3js.org/d3.v7.min.js"></script>
 
     <title>Grimdark Stats</title>
   </head>
@@ -14,17 +13,26 @@
       <h1>Welcome to Grimdark stats!</h1>
     </header>
 
-		<div class="content">
-			<section class="side-bar">
-				<p>Please click on each major faction to see their subfactions! <br>
-				Hover over each subfaction to see their win rate!</p>
-			</section>
+    <div class="content">
+      <section class="side-bar">
+        <p>
+          Please click on each major faction to see their subfactions! <br />
+          Hover over each subfaction to see their win rate!
+        </p>
+				<div class="mc-faction">
+					Most Common Faction Played 
+				</div>
 
-			<section class="graph">
-				<header>Most Common Faction Played</header>
-				<svg></svg>
-			</section>
-		</div>
+				<div class="wr-faction">
+					Faction Win Rates
+				</div>
+      </section>
+
+      <section class="graph-wrapper">
+        <section class="graph">
+        </section>
+      </section>
+    </div>
     <script src="./dist/main.js"></script>
   </body>
 </html>
