+++ src/styles/index.css
@@ -1,26 +1,84 @@
 rect {
-	fill: #003d26
+  fill: #003d26;
 }
 body {
-	margin: 0 32px;
+  margin: 0 32px;
 }
 text {
-	font-size: 1.4em;
+  font-size: 1.4em;
 }
 
 .slice:hover {
-	opacity: .50;
+  opacity: 0.5;
 }
 
 .content {
-	display: flex;
-
+  display: flex;
 }
 
 .side-bar {
-	width: 20vw;
+  width: 20vw;
 }
 
 header {
-	margin-left: 60px;
+}
+
+.graph {
+  display: flex;
+  flex-wrap: wrap;
+	max-width: 70vw;
+}
+/* 
+.round {
+  transform: rotate(-90deg);
+  transition: all 1s ease-in-out;
+
+  fill: none;
+  stroke-width: 8;
+  stroke-linecap: round;
+  stroke: red;
+} */
+svg {
+	transition: all 300ms ease;
+	margin: 10px;
+}
+
+
+.round .background {
+  fill: #e6e7e8;
+}
+
+.circle:hover {
+	transform: scale(1.05);
+}
+
+.round text.win-percentage {
+  font-family: Graphik;
+  font-size: 28px;
+  letter-spacing: -0.03em;
+}
+.round text.description {
+  font-family: Graphik;
+  font-size: 12px;
+  fill: #5a5a5a;
+}
+
+.round text.faction-title {
+	font-family: Graphik;
+	font-size: 16px;
+	fill: grey;
+}
+.mc-faction:hover {
+	cursor: pointer;
+}
+
+.wr-faction:hover {
+	cursor: pointer;
+}
+
+
+.sunburst {
+	position: absolute;
+	top: 0;
+	left: 400px;
 }
